shopping-cart-analysis-python/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alchemy==20.5
alembic==1.13.1
altair==5.2.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.3.0
Automat==25.4.16
Babel==2.14.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.7.0
boto3==1.40.11
botocore==1.40.11
cachetools==5.3.2
camelot-py==0.11.0
certifi==2025.6.15
cffi==1.17.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.2.1
cloudpickle==3.0.0
comm==0.2.1
constantly==23.10.4
contourpy==1.2.0
cryptography==45.0.4
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
Django==5.0.3
dm-tree==0.1.8
dnspython==2.7.0
docutils==0.22
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.2.0
factory_boy==3.3.3
Faker==23.2.1
fastjsonschema==2.19.1
favicon==0.7.0
filelock==3.18.0
flatbuffers==23.5.26
fonttools==4.49.0
fqdn==1.5.1
frozendict==2.4.0
frozenlist==1.4.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
htbuilder==0.6.2
httpcore==1.0.4
httpx==0.27.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.10
importlib-metadata==7.0.1
incremental==24.7.2
ipykernel==6.29.3
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==5.13.2
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
json5==0.9.17
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.12.0
jupyter-lsp==2.2.3
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.0.5
kiwisolver==1.4.5
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langsmith==0.1.8
libclang==16.0.6
lockfile==0.12.2
lxml==6.0.0
Mako==1.3.2
Markdown==3.8.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
mistune==3.0.2
ml-dtypes==0.3.2
more-itertools==10.2.0
multidict==6.0.5
munch==4.0.0
mypy-extensions==1.0.0
namex==0.0.7
narwhals==2.1.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.3
pdfminer.six==20231228
pdfplumber==0.11.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==6.3.0
prometheus_client==0.22.1
prompt-toolkit==3.0.43
propcache==0.3.2
Protego==0.4.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.8.1b0
PyDispatcher==2.0.7
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.7
pymongo==4.14.0
pyOpenSSL==25.1.0
pyparsing==3.1.1
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
python-daemon==2.2.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-engineio==4.12.1
python-json-logger==2.0.7
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.5.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==25.1.2
queuelib==1.8.0
redis==6.2.0
referencing==0.33.0
requests==2.27.1
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.14.7
rpds-py==0.18.0
s3transfer==0.13.1
Scrapy==2.13.0
seaborn==0.13.2
selenium==4.32.0
Send2Trash==1.8.2
service-identity==24.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.17.1
snowflake-snowpark-python==1.36.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.4.4
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.31.1
streamlit-aggrid==0.3.4.post3
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
TA-Lib==0.4.32
tabula-py==2.9.3
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
tinycss2==1.2.1
tldextract==5.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.3
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
typer==0.16.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.20
uv==0.7.9
uvloop==0.21.0
validators==0.22.0
w3lib==2.3.1
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope.interface==7.2
